server/image-service/requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.0
sqlalchemy>=2.0.40
psycopg2-binary>=2.9.10
alembic==1.15.2
python-jose==3.4.0
python-multipart>=0.0.20
pydantic>=2.11.3
pydantic-settings>=2.0.0
email-validator>=2.0.0
pillow==11.1.0
python-dotenv>=1.1.0
httpx>=0.28.1
asyncpg>=0.30.0
cloudinary>=1.44.0